# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Finance Tracker API
  description: Finance Data Processing and Access Control Backend API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Record:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [INCOME, EXPENSE]
        amount:
          type: number
        category:
          type: string
        date:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/records:
    get:
      tags:
        - Financial Records
      summary: List financial records
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Record'
                  meta:
                    type: object
    post:
      tags:
        - Financial Records
      summary: Create a new financial record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, amount, category, date]
              properties:
                type:
                  type: string
                  enum: [INCOME, EXPENSE]
                amount:
                  type: number
                category:
                  type: string
                date:
                  type: string
                  format: date
                notes:
                  type: string
      responses:
        '201':
          description: Record created
...